+++ src/odin/narrative_dictionary.cc
@@ -340,6 +340,10 @@ void NarrativeDictionary::Load(const boost::property_tree::ptree& narrative_pt) 
   LOG_TRACE("Populate level_change_subset");
   // Populate level_change_subset
   Load(level_change_subset, narrative_pt.get_child(kLevelChangeKey));
+
+  LOG_TRACE("Populate lbike_share_verbal_subset");
+  // Populate bike_share_verbal_subset
+  Load(bike_share_verbal, narrative_pt.get_child(kBikeShareVerbalKey));
 }
 
 void NarrativeDictionary::Load(PhraseSet& phrase_handle,