+++ test/astar_bikeshare.cc
@@ -51,13 +51,10 @@ struct route_tester {
   vo::odin_worker_t odin_worker;
 };
 
-using BssManeuverType = valhalla::DirectionsLeg_Maneuver_BssManeuverType;
-
 void test_request(const std::string& request,
                   const std::vector<TravelMode>& expected_travel_modes,
                   const std::vector<std::string>& expected_route,
-                  // We mark only the maneuvers that are RentBike and ReturnBike
-                  const std::map<size_t, BssManeuverType>& expected_bss_maneuver,
+                  const std::map<size_t, DirectionsLeg_Maneuver_Type>& expected_maneuver_types = {},
                   const std::map<size_t, std::string>& expected_bss_ref = {},
                   const std::optional<std::string>& expected_shape = {}) {
 
@@ -82,7 +79,6 @@ void test_request(const std::string& request,
   std::vector<TravelMode> travel_modes;
 
   std::vector<std::string> route;
-
   for (const auto& d : directions) {
     if (expected_shape) {
       EXPECT_TRUE(test::encoded_shape_equality(d.shape(), *expected_shape))
@@ -90,20 +86,16 @@ void test_request(const std::string& request,
     }
     size_t idx = -1;
     for (const auto& m : d.maneuver()) {
-      auto it = expected_bss_maneuver.find(++idx);
-      if (it == expected_bss_maneuver.end()) {
-        EXPECT_EQ(m.bss_maneuver_type(),
-                  BssManeuverType::DirectionsLeg_Maneuver_BssManeuverType_kNoneAction)
-            << "BSS maneuver type at " + std::to_string(idx) + " is incorrect";
-
-      } else {
-        EXPECT_EQ(m.bss_maneuver_type(), it->second)
-            << "BSS maneuver type at " + std::to_string(idx) + " is incorrect";
-      }
+      ++idx;
       auto search = expected_bss_ref.find(idx);
       if (search != expected_bss_ref.end()) {
         EXPECT_EQ(m.bss_info().ref(), search->second)
             << "bss_info.osm_node_id at " + std::to_string(idx) + " is incorrect";
+      }
+      if (!expected_maneuver_types.empty()) {
+        auto it_maneuver_type = expected_maneuver_types.find(idx);
+        EXPECT_EQ(m.type(), it_maneuver_type->second)
+            << "maneuver type at " + std::to_string(idx) + " is incorrect";
       }
       travel_modes.push_back(m.travel_mode());
       std::string name;
@@ -140,23 +132,39 @@ void test_request(const std::string& request,
 TEST(AstarBss, test_With_Mode_Changes) {
   std::string request =
       R"({"locations":[{"lat":48.864655,"lon":2.361374},{"lat":48.859608,"lon":2.36117}],"costing":"bikeshare",
-	       "costing_options":{"pedestrian":{"bss_rent_cost":0,"bss_rent_penalty":0},
-	                          "bicycle"   :{"bss_return_cost":0,"bss_return_penalty":0}}})";
+           "costing_options":{"pedestrian":{"bss_rent_cost":0,"bss_rent_penalty":0},
+                              "bicycle"   :{"bss_return_cost":0,"bss_return_penalty":0}}})";
+
   std::vector<TravelMode> expected_travel_modes{TravelMode::kPedestrian, TravelMode::kBicycle,
                                                 TravelMode::kPedestrian};
 
   std::vector<std::string> expected_route{"Rue Gabriel Vicaire", "Rue Perrée",
-                                          "Rue Perrée",          "Rue Caffarelli",
-                                          "Rue de Bretagne",     "Rue de Turenne",
-                                          "Rue du Parc Royal",   "Place de Thorigny",
+                                          "Rue Perrée",          "Rue Perrée",
+                                          "Rue Caffarelli",      "Rue de Bretagne",
+                                          "Rue de Turenne",      "Rue du Parc Royal",
+                                          "Place de Thorigny",   "Rue de la Perle",
                                           "Rue de la Perle",     "Rue de la Perle"};
 
-  const std::map<size_t, BssManeuverType>&
-      expected_bss_maneuver{{2, DirectionsLeg_Maneuver_BssManeuverType_kRentBikeAtBikeShare},
-                            {9, DirectionsLeg_Maneuver_BssManeuverType_kReturnBikeAtBikeShare}};
-  const std::map<size_t, std::string>& expected_bss_ref{{2, "3006"}, {9, "3008"}};
+  const std::map<size_t, DirectionsLeg_Maneuver_Type>
+      expected_maneuver_types{{0, DirectionsLeg_Maneuver_Type_kStart},
+                              {1, DirectionsLeg_Maneuver_Type_kLeft},
+                              {2, DirectionsLeg_Maneuver_Type_kRentBikeAtBikeShare},
+                              {3, DirectionsLeg_Maneuver_Type_kRight},
+                              {4, DirectionsLeg_Maneuver_Type_kRight},
+                              {5, DirectionsLeg_Maneuver_Type_kLeft},
+                              {6, DirectionsLeg_Maneuver_Type_kRight},
+                              {7, DirectionsLeg_Maneuver_Type_kRight},
+                              {8, DirectionsLeg_Maneuver_Type_kSlightRight},
+                              {9, DirectionsLeg_Maneuver_Type_kSlightLeft},
+                              {10, DirectionsLeg_Maneuver_Type_kReturnBikeAtBikeShare},
+                              {11, DirectionsLeg_Maneuver_Type_kUturnRight},
+                              {12, DirectionsLeg_Maneuver_Type_kDestinationRight}};
+  const std::map<size_t, std::string>& expected_bss_ref{{2, "3006"},
+                                                        {2, "3006"},
+                                                        {10, "3008"},
+                                                        {11, "3008"}};
 
-  test_request(request, expected_travel_modes, expected_route, expected_bss_maneuver,
+  test_request(request, expected_travel_modes, expected_route, expected_maneuver_types,
                expected_bss_ref);
 }
 
@@ -168,9 +176,9 @@ TEST(AstarBss, test_BSS_mode_Without_Mode_Changes) {
 
   std::string request =
       R"({"locations":[{"lat":48.865020,"lon":2.369113},{"lat":48.859782,"lon":2.36101}],
-	       "costing":"bikeshare",
-	       "costing_options":{"pedestrian":{"bss_rent_cost":0,"bss_rent_penalty":1800},
-	                          "bicycle"   :{"bss_return_cost":0,"bss_return_penalty":1800}}})";
+           "costing":"bikeshare",
+           "costing_options":{"pedestrian":{"bss_rent_cost":0,"bss_rent_penalty":1800},
+                              "bicycle"   :{"bss_return_cost":0,"bss_return_penalty":1800}}})";
 
   std::vector<TravelMode> expected_travel_modes{TravelMode::kPedestrian};
 
@@ -181,9 +189,7 @@ TEST(AstarBss, test_BSS_mode_Without_Mode_Changes) {
                                           "Rue de Turenne",       "Rue Vieille du Temple",
                                           "Rue de la Perle"};
 
-  const std::map<size_t, BssManeuverType>& expected_bss_maneuver{};
-
-  test_request(request, expected_travel_modes, expected_route, expected_bss_maneuver);
+  test_request(request, expected_travel_modes, expected_route);
 }
 
 /*
@@ -193,9 +199,9 @@ TEST(AstarBss, test_BSS_mode_Without_Mode_Changes) {
 TEST(AstarBss, test_BSS_mode_Without_Mode_Changes_2) {
   std::string request =
       R"({"locations":[{"lat":48.865020,"lon":2.369113},{"lat":48.859782,"lon":2.36101}],
-	       "costing":"bikeshare",
-	       "costing_options":{"pedestrian":{"bss_rent_cost":1800,"bss_rent_penalty":0},
-	                          "bicycle"   :{"bss_return_cost":1800,"bss_return_penalty":0}}})";
+           "costing":"bikeshare",
+           "costing_options":{"pedestrian":{"bss_rent_cost":1800,"bss_rent_penalty":0},
+                              "bicycle"   :{"bss_return_cost":1800,"bss_return_penalty":0}}})";
 
   std::vector<TravelMode> expected_travel_modes{TravelMode::kPedestrian};
 
@@ -206,20 +212,19 @@ TEST(AstarBss, test_BSS_mode_Without_Mode_Changes_2) {
                                           "Rue de Turenne",       "Rue Vieille du Temple",
                                           "Rue de la Perle"};
 
-  const std::map<size_t, BssManeuverType>& expected_bss_maneuver{};
-
-  test_request(request, expected_travel_modes, expected_route, expected_bss_maneuver);
+  test_request(request, expected_travel_modes, expected_route);
 }
 
 // We test if the bss connection edges respect the forward/reverse access
 TEST(AstarBss, test_With_Mode_Changes_2) {
   std::string request =
       R"({"locations":[{"lat":48.8601411,"lon":2.3716413},{"lat":48.8594916,"lon":2.3602581}],"costing":"bikeshare",
-	       "costing_options":{"pedestrian":{"bss_rent_cost":0,"bss_rent_penalty":0},
-	                          "bicycle"   :{"bss_return_cost":0,"bss_return_penalty":0}}})";
+           "costing_options":{"pedestrian":{"bss_rent_cost":0,"bss_rent_penalty":0},
+                              "bicycle"   :{"bss_return_cost":0,"bss_return_penalty":0}}})";
   std::vector<TravelMode> expected_travel_modes{TravelMode::kPedestrian, TravelMode::kBicycle,
                                                 TravelMode::kPedestrian};
   std::vector<std::string> expected_route{"Rue Pelée",
+                                          "Rue Pelée",
                                           "Rue Pelée",
                                           "Rue Alphonse Baudin",
                                           "Rue Saint-Sébastien",
@@ -230,13 +235,30 @@ TEST(AstarBss, test_With_Mode_Changes_2) {
                                           "Place de Thorigny",
                                           "Rue de la Perle",
                                           "Rue de la Perle",
+                                          "Rue de la Perle",
                                           "Rue Vieille du Temple"};
-  const std::map<size_t, BssManeuverType>&
-      expected_bss_maneuver{{1, DirectionsLeg_Maneuver_BssManeuverType_kRentBikeAtBikeShare},
-                            {10, DirectionsLeg_Maneuver_BssManeuverType_kReturnBikeAtBikeShare}};
-  const std::map<size_t, std::string>& expected_bss_ref{{1, "11103"}, {10, "3008"}};
+  const std::map<size_t, DirectionsLeg_Maneuver_Type>&
+      expected_maneuver_types{{0, DirectionsLeg_Maneuver_Type_kStart},
+                              {1, DirectionsLeg_Maneuver_Type_kRentBikeAtBikeShare},
+                              {2, DirectionsLeg_Maneuver_Type_kSlightLeft},
+                              {3, DirectionsLeg_Maneuver_Type_kRight},
+                              {4, DirectionsLeg_Maneuver_Type_kLeft},
+                              {5, DirectionsLeg_Maneuver_Type_kRight},
+                              {6, DirectionsLeg_Maneuver_Type_kLeft},
+                              {7, DirectionsLeg_Maneuver_Type_kLeft},
+                              {8, DirectionsLeg_Maneuver_Type_kRight},
+                              {9, DirectionsLeg_Maneuver_Type_kSlightRight},
+                              {10, DirectionsLeg_Maneuver_Type_kSlightLeft},
+                              {11, DirectionsLeg_Maneuver_Type_kReturnBikeAtBikeShare},
+                              {12, DirectionsLeg_Maneuver_Type_kLeft},
+                              {13, DirectionsLeg_Maneuver_Type_kLeft},
+                              {14, DirectionsLeg_Maneuver_Type_kDestinationLeft}};
+  const std::map<size_t, std::string>& expected_bss_ref{{1, "11103"},
+                                                        {2, "11103"},
+                                                        {11, "3008"},
+                                                        {12, "3008"}};
 
-  test_request(request, expected_travel_modes, expected_route, expected_bss_maneuver,
+  test_request(request, expected_travel_modes, expected_route, expected_maneuver_types,
                expected_bss_ref);
 }
 
@@ -247,10 +269,7 @@ TEST(AstarBss, test_Pedestrian) {
   std::vector<TravelMode> expected_travel_modes{TravelMode::kPedestrian};
   std::vector<std::string> expected_route{"Rue de la Perle", "Rue Vieille du Temple", "Rue Froissart",
                                           "Rue Commines", "Rue Amelot"};
-  // There shouldn't be any bss maneuvers
-  const std::map<size_t, BssManeuverType>& expected_bss_maneuver{};
-
-  test_request(request, expected_travel_modes, expected_route, expected_bss_maneuver);
+  test_request(request, expected_travel_modes, expected_route);
 }
 
 // When bicycle is chosen as travel_mode, the departure edge must NOT be a bss connections edge
@@ -260,10 +279,7 @@ TEST(AstarBss, test_Bicycle) {
   std::vector<TravelMode> expected_travel_modes{TravelMode::kBicycle};
   std::vector<std::string> expected_route{"Rue de la Perle", "Rue des Archives", "Rue de Bretagne",
                                           "Rue Commines", "Rue Amelot"};
-  // There shouldn't be any bss maneuvers
-  const std::map<size_t, BssManeuverType>& expected_bss_maneuver{};
-
-  test_request(request, expected_travel_modes, expected_route, expected_bss_maneuver);
+  test_request(request, expected_travel_modes, expected_route);
 }
 
 // When auto is chosen as travel_mode, the departure edge must NOT be a bss connections edge
@@ -273,10 +289,7 @@ TEST(AstarBss, test_Auto) {
   std::vector<TravelMode> expected_travel_modes{TravelMode::kDrive};
   std::vector<std::string> expected_route{"Rue de la Perle", "Rue des Archives", "Rue Pastourelle",
                                           "Rue de Turenne",  "Rue Commines",     "Rue Amelot"};
-  // There shouldn't be any bss maneuvers
-  const std::map<size_t, BssManeuverType>& expected_bss_maneuver{};
-
-  test_request(request, expected_travel_modes, expected_route, expected_bss_maneuver);
+  test_request(request, expected_travel_modes, expected_route);
 }
 
 // When auto is chosen as travel_mode, the departure edge must NOT be a bss connections edge
@@ -288,10 +301,7 @@ TEST(AstarBss, test_Truck) {
       expected_route{"Rue de la Perle",     "Rue des Archives",       "Rue Pastourelle",
                      "Rue du Temple",       "Place de la République", "Place de la République",
                      "Boulevard du Temple", "Rue Oberkampf",          "Rue Amelot"};
-  // There shouldn't be any bss maneuvers
-  const std::map<size_t, BssManeuverType>& expected_bss_maneuver{};
-
-  test_request(request, expected_travel_modes, expected_route, expected_bss_maneuver);
+  test_request(request, expected_travel_modes, expected_route);
 }
 
 // In this test case, the bike share station(48.8690345, 2.3622890) is located a dedicated cyclelane
@@ -317,6 +327,7 @@ TEST(AstarBss, test_BSSConnections_on_Pedestrian_and_Bicycle) {
   std::vector<std::string> expected_route{"Rue Perrée",
                                           "Rue Perrée",
                                           "Rue Perrée",
+                                          "Rue Perrée",
                                           "Rue Eugène Spuller",
                                           "Rue Béranger",
                                           "Rue du Temple",
@@ -324,18 +335,36 @@ TEST(AstarBss, test_BSSConnections_on_Pedestrian_and_Bicycle) {
                                           "Place de la République",
                                           "Boulevard de Magenta",
                                           "Rue du Château d'Eau",
+                                          "Rue du Château d'Eau",
                                           "Rue du Château d'Eau"};
 
-  const std::map<size_t, BssManeuverType>&
-      expected_bss_maneuver{{2, DirectionsLeg_Maneuver_BssManeuverType_kRentBikeAtBikeShare},
-                            {11, DirectionsLeg_Maneuver_BssManeuverType_kReturnBikeAtBikeShare}};
+  const std::map<size_t, std::string>& expected_bss_ref{{2, "3006"},
+                                                        {3, "3006"},
+                                                        {12, "10011"},
+                                                        {13, "10011"}};
 
-  const std::map<size_t, std::string>& expected_bss_ref{{2, "3006"}, {11, "10011"}};
+  const std::map<size_t, DirectionsLeg_Maneuver_Type>&
+      expected_maneuver_types{{0, DirectionsLeg_Maneuver_Type_kStart},
+                              {1, DirectionsLeg_Maneuver_Type_kSlightRight},
+                              {2, DirectionsLeg_Maneuver_Type_kRentBikeAtBikeShare},
+                              {3, DirectionsLeg_Maneuver_Type_kUturnRight},
+                              {3, DirectionsLeg_Maneuver_Type_kRight},
+                              {4, DirectionsLeg_Maneuver_Type_kLeft},
+                              {5, DirectionsLeg_Maneuver_Type_kLeft},
+                              {6, DirectionsLeg_Maneuver_Type_kRight},
+                              {7, DirectionsLeg_Maneuver_Type_kLeft},
+                              {8, DirectionsLeg_Maneuver_Type_kRight},
+                              {9, DirectionsLeg_Maneuver_Type_kContinue},
+                              {10, DirectionsLeg_Maneuver_Type_kLeft},
+                              {11, DirectionsLeg_Maneuver_Type_kSlightRight},
+                              {12, DirectionsLeg_Maneuver_Type_kReturnBikeAtBikeShare},
+                              {13, DirectionsLeg_Maneuver_Type_kContinue},
+                              {14, DirectionsLeg_Maneuver_Type_kDestination}};
 
   std::string expected_shape =
       "e~le|A_ldoCyD~IoAtCkArC]z@kBpEeAsAdArAjBqE\\{@jAsCad@ai@yAgBo@iCuF_Ua@_B[uAyQgz@i@cCwAt@mg@bXyt@b`@yCvAyBqH{EgLiCvEoD|G{\\`r@wFqHoPqTy@gAyAkBe@o@i@q@{D_CeB{@wCfC{XfVt@jCjA~Dn@xB?lBcA|BV\\f@r@wBlE";
 
-  test_request(request, expected_travel_modes, expected_route, expected_bss_maneuver,
+  test_request(request, expected_travel_modes, expected_route, expected_maneuver_types,
                expected_bss_ref, expected_shape);
 }